Scene::Load moved to a separate file to parallelize compilation a bit. #10
